commit
I refactored the grid of the book list to have the list on the left and the CommentsArea always visible on the right. I also created the initial state for the CommentsArea for when there's no data to be displayed. The CommentsArea behavior: every time the user clicks on a card, a fetch should be performed and comments should update. I passed the single ID property and let the Component handle the fetch by using the componentDidUpdate method.