I used React Hooks on my old project to convert every Class Component into a Functional one.
I used useState() and useEffect() to mimic Class Component capabilities(having in mind that useEffect() can be used like componentDidMount as well as componentDidUpdate, depending on what I put on its dependency array. I also maintained the values for the input fields in AddComment in a single object, using just one useState() for the whole form.